Regenerate the yarn.lock file
This doesn't change any dependency versions required for consuming repos,
but addresses some the dependency CVEs in this repo.

(Some devDependency CVEs remain coming in through storybook versions, but
this project is not widely used at Culture Amp anymore and the effort to upgrade
to new storybook versions likely isn't worth it).